# openapi/openapi.yaml
openapi: 3.1.2
info:
  title: ocservia API
  version: 1.0.0
  description: Public control-plane API contract.
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
servers:
  - url: /api/v1
security:
  - oidc: []
  - bearerAuth: []
...
paths:
...
  /operations:
    get:
      operationId: listOperations
      summary: List operations
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Operations ordered from newest to oldest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /operations/{operation_id}:
    get:
      operationId: getOperation
      summary: Get an asynchronous operation
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /events:
    get:
      operationId: listEvents
      summary: Rebuild platform events from durable state
      tags: [events]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: after
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UuidV7"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Events ordered from oldest to newest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformEventPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    oidc:
      type: apiKey
      in: cookie
      name: __Host-ocservia_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Secure, HttpOnly, SameSite server-side session established by OIDC Authorization Code plus PKCE S256.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous page.
      required: false
      schema:
        $ref: "#/components/schemas/Cursor"
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    OperationId:
      name: operation_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UuidV7"
...
    WorkspaceId:
      name: X-Workspace-ID
      in: header
      description: Required when the principal can access more than one workspace; otherwise the sole authorized workspace is inferred.
      required: false
      schema: { $ref: "#/components/schemas/UuidV7" }
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    UuidV7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      example: 0193c1b2-7d8a-7f01-8a2d-1e5b4f302901
    UtcDateTime:
      type: string
      format: date-time
      description: RFC 3339 timestamp normalized to UTC.
      example: "2026-08-01T16:00:00Z"
...
    Cursor:
      type: string
      minLength: 1
      maxLength: 2048
      description: Opaque pagination position; clients must not parse it.
    OperationState:
      type: string
      enum:
        - draft
        - queued
        - dispatched
        - accepted
        - running
        - succeeded
        - failed
        - unknown
        - expired
        - rolled_back
        - offline_pending
        - drifted
        - superseded
    Operation:
      type: object
      additionalProperties: false
      required: [id, state, version, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UuidV7"
        state:
          $ref: "#/components/schemas/OperationState"
        node_id:
          $ref: "#/components/schemas/UuidV7"
        command_id:
          $ref: "#/components/schemas/UuidV7"
        config_apply_state:
          type: string
          enum: [queued, dispatched, accepted, running, succeeded, failed, rolled_back, failed_critical, unknown, expired]
          description: Typed configuration outcome when this operation applies a configuration plan.
        config_apply_failure_code:
          type: string
          minLength: 1
          maxLength: 128
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          $ref: "#/components/schemas/UtcDateTime"
        updated_at:
          $ref: "#/components/schemas/UtcDateTime"
        expires_at:
          $ref: "#/components/schemas/UtcDateTime"
...
    PlatformEventType:
      type: string
      enum: [connected, disconnected, command_result, simulation_result, heartbeat, error, path_changed, telemetry]
    Traceparent:
      type: string
      pattern: "^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$"
    PlatformEvent:
      type: object
      additionalProperties: false
      required: [id, node_id, type, traceparent, occurred_at]
      properties:
        id:
          $ref: "#/components/schemas/UuidV7"
        node_id:
          $ref: "#/components/schemas/UuidV7"
        type:
          $ref: "#/components/schemas/PlatformEventType"
        traceparent:
          $ref: "#/components/schemas/Traceparent"
        occurred_at:
          $ref: "#/components/schemas/UtcDateTime"
    PlatformEventPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/PlatformEvent"
        page:
          $ref: "#/components/schemas/PageInfo"
    PageInfo:
      type: object
      additionalProperties: false
      required: [has_more]
      properties:
        has_more:
          type: boolean
        next_cursor:
          $ref: "#/components/schemas/Cursor"
    OperationPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/Operation"
        page:
          $ref: "#/components/schemas/PageInfo"
...
    Problem:
      type: object
      additionalProperties: true
      description: RFC 9457 Problem Details response.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4096
        instance:
          type: string
          format: uri-reference
        trace_id:
          type: string
          pattern: "^[0-9a-f]{32}$"